requirements.txt
cosyvoice==0.0.8
dashscope==1.24.1
decord==0.6.0
diffusers==0.34.0
easydict==1.13
einops==0.8.1
flash_attn==2.7.4.post1
ftfy==6.3.1
GitPython==3.1.45
hydra-core==1.3.2
imageio==2.37.0
librosa==0.11.0
loguru==0.7.3
matplotlib==3.10.7
modelscope==1.29.2
moviepy==2.2.1
numpy==2.3.5
omegaconf==2.3.0
onnxruntime==1.23.2
opencv_python==4.11.0.86
packaging==25.0
peft==0.18.0
Pillow==12.0.0
regex==2025.7.34
Requests==2.32.5
safetensors==0.7.0
sam2==1.1.0
scipy==1.16.3
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.55.0
